Netting UI fixes & refactor (#816)
* refactor: prettier autoformatting

* typo: change the tooltip copy

* fix: put back price impact color

* refactor: split crab into deposit and withdraw components

* style: make the tradeSection sticky + cleanup

* feat: debounce input values

* fix: show disabled button on error

* refactor: put common style in the same place

* typo: change "queued" to "initiated"

* fix: fix button styling

* refactor: use queueOptionAvailable and useQueue instead

* style: remove sticky behaviour

* typo: change wording

* fix: show queueNotice at all times when queue is being used

* fix: price impact should only have red and grey colors

* refactor: put crab metrics into its own component

* fix: show average auction price impact in case queue is being used

* fix: refresh usdcQueue and crabQueue values on tx confirmation

* feat: add price impact tooltip for standard option

* styles: match rounded button styling + fix button state
