+++ packages/frontend/package.json
@@ -69,6 +69,7 @@
         "react-use": "^17.3.2",
         "set-interval-async": "^2.0.3",
         "subscriptions-transport-ws": "^0.11.0",
+        "use-debounce": "^9.0.2",
         "web3": "^1.5.2"
     },
     "devDependencies": {
