+++ packages/frontend/src/components/Button/index.tsx
@@ -94,10 +94,16 @@ export const PrimaryButtonNew = withStyles((theme) => ({
 
 export const RoundedButton = withStyles((theme) => ({
   root: {
+    backgroundColor: theme.palette.background.stone,
     '&:disabled': {
-      color: 'rgba(255, 255, 255, 0.4)',
-      backgroundColor: theme.palette.background.lightStone,
+      color: 'rgba(255, 255, 255, 0.1)',
+      backgroundColor: 'rgba(36, 39, 40, 0.6)',
+      opacity: 1,
     },
     textTransform: 'initial',
+    borderRadius: '20px',
+    padding: '4px 24px',
+    fontSize: '16px',
+    fontWeight: 700,
   },
 }))(Button)
