+++ packages/frontend/src/components/Metric.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { Typography, Box, BoxProps, Tooltip } from '@material-ui/core'
 import { makeStyles, createStyles } from '@material-ui/core/styles'
 import InfoIcon from '@material-ui/icons/InfoOutlined'
+import clsx from 'clsx'
 
 interface StyleProps {
   isSmall: boolean
@@ -27,22 +28,33 @@ const useStyles = makeStyles((theme) =>
       width: 'max-content',
       fontFamily: 'DM Mono',
     },
+    success: {
+      color: theme.palette.success.main,
+    },
+    error: {
+      color: theme.palette.error.main,
+    },
   }),
 )
 
 interface MetricProps {
   label: string | React.ReactNode
   value: string | React.ReactNode
+  textColor?: 'success' | 'error'
   isSmall?: boolean
 }
 
-const Metric: React.FC<MetricProps & BoxProps> = ({ label, value, isSmall = false, ...props }) => {
+const Metric: React.FC<MetricProps & BoxProps> = ({ label, value, isSmall = false, textColor, ...props }) => {
   const classes = useStyles({ isSmall })
 
   return (
     <Box className={classes.container} flex="1" display="flex" flexDirection="column" {...props}>
       {typeof label === 'string' ? <Typography className={classes.label}>{label}</Typography> : label}
-      {typeof value === 'string' ? <Typography className={classes.value}>{value}</Typography> : value}
+      {typeof value === 'string' ? (
+        <Typography className={clsx(classes.value, textColor && classes[textColor])}>{value}</Typography>
+      ) : (
+        value
+      )}
     </Box>
   )
 }
@@ -68,18 +80,17 @@ const useLabelStyles = makeStyles((theme) =>
   }),
 )
 
-export const MetricLabel: React.FC<{ label: string | React.ReactNode; tooltipTitle: string | React.ReactNode }> = ({
-  label,
-  tooltipTitle,
-}) => {
+export const MetricLabel: React.FC<{ label: string; tooltipTitle?: string }> = ({ label, tooltipTitle }) => {
   const classes = useLabelStyles()
 
   return (
     <div className={classes.labelContainer}>
       <Typography className={classes.label}>{label}</Typography>
-      <Tooltip title={tooltipTitle || ''}>
-        <InfoIcon fontSize="small" className={classes.infoIcon} />
-      </Tooltip>
+      {tooltipTitle ? (
+        <Tooltip title={tooltipTitle}>
+          <InfoIcon fontSize="small" className={classes.infoIcon} />
+        </Tooltip>
+      ) : null}
     </div>
   )
 }
