+++ packages/frontend/src/components/Strategies/Crab/CrabPositionV2.tsx
@@ -11,7 +11,7 @@ import { useCrabPositionV2 } from '@hooks/useCrabPosition/useCrabPosition'
 import Metric, { MetricLabel } from '@components/Metric'
 import { formatCurrency, formatNumber } from '@utils/formatter'
 import { pnlInPerctv2 } from 'src/lib/pnl'
-import { crabQueuedAtom, crabUSDValueAtom, usdcQueuedAtom } from 'src/state/crab/atoms'
+import { crabQueuedAtom, crabUSDValueAtom, usdcQueuedAtom } from '@state/crab/atoms'
 import { toTokenAmount } from '@utils/calculations'
 import { BIG_ZERO, USDC_DECIMALS } from '@constants/index'
 import { useTransactionStatus } from '@state/wallet/hooks'
