+++ packages/frontend/src/constants/index.ts
@@ -21,6 +21,8 @@ export const VOL_PERCENT_FIXED = 0.1
 
 export const VOL_PERCENT_SCALAR = 0.1
 
+export const AVERAGE_AUCTION_PRICE_IMPACT = 0.79
+
 //7min, 420 seconds
 export const TWAP_PERIOD = 420
 
@@ -35,6 +37,7 @@ export const INDEX_SCALE = 10000
 export const OSQUEETH_DECIMALS = 18
 export const WETH_DECIMALS = 18
 export const USDC_DECIMALS = 6
+export const CRAB_TOKEN_DECIMALS = 18
 
 export const SWAP_EVENT_TOPIC = '0xc42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67'
